requirements.txt
openai
python-dotenv
streamlit
gitpython
requests
flask